+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/metric/makers/BaseProtocolMetricMaker.java
@@ -8,10 +8,12 @@ import com.yahoo.bard.webservice.data.metric.MetricDictionary;
 import com.yahoo.bard.webservice.data.metric.TemplateDruidQuery;
 import com.yahoo.bard.webservice.data.metric.mappers.ResultSetMapper;
 import com.yahoo.bard.webservice.data.metric.protocol.DefaultSystemMetricProtocols;
+import com.yahoo.bard.webservice.data.metric.protocol.GeneratedMetricInfo;
 import com.yahoo.bard.webservice.data.metric.protocol.ProtocolMetric;
 import com.yahoo.bard.webservice.data.metric.protocol.ProtocolSupport;
 import com.yahoo.bard.webservice.data.metric.protocol.ProtocolMetricImpl;
 
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -59,10 +61,66 @@ public abstract class BaseProtocolMetricMaker extends MetricMaker implements Mak
             LogicalMetricInfo logicalMetricInfo,
             List<LogicalMetric> dependentMetrics
     ) {
-        TemplateDruidQuery partialQuery = makePartialQuery(logicalMetricInfo, dependentMetrics);
-        ResultSetMapper calculation = makeCalculation(logicalMetricInfo, dependentMetrics);
-        ProtocolSupport protocolSupport = makeProtocolSupport(logicalMetricInfo, dependentMetrics);
+        LogicalMetric dependentMetric = dependentMetrics.get(0);
+        LogicalMetric renamedDependentMetric = renameIfConflicting(logicalMetricInfo.getName(), dependentMetric);
+        List<LogicalMetric> renamedDependentMetrics =
+                dependentMetrics.size() == 1 ? Collections.singletonList(renamedDependentMetric) : dependentMetrics;
+        TemplateDruidQuery partialQuery = makePartialQuery(logicalMetricInfo, renamedDependentMetrics);
+        ResultSetMapper calculation = makeCalculation(logicalMetricInfo, renamedDependentMetrics);
+        ProtocolSupport protocolSupport = makeProtocolSupport(logicalMetricInfo, renamedDependentMetrics);
         return new ProtocolMetricImpl(logicalMetricInfo, partialQuery, calculation, protocolSupport);
+    }
+
+    /**
+     * Renames the provided logical metric if there is a name conflict between it and the final desired output name.
+     * If there is no name conflict the input metric is returned unchanged. Otherwise, the input metric is renamed to
+     * not conflict with the final output name.
+     * <p>
+     * This method specifically the {@link GeneratedMetricInfo} typing on the input metric if it has to be renamed. All
+     * other subclasses of {@link LogicalMetricInfo} are lost during the renaming. If maintaining more LogicalMetricInfo
+     * types is required, a withName method must «should be added to the LogicalMetricInfo interface.
+     *
+     * @param finalOutputName  The name to check for conflicts on
+     * @param dependentMetric  The dependent metric to check for a conflicting name
+     * @return  The metric without a name conflicting with finalOutputName
+     */
+     protected LogicalMetric renameIfConflicting(String finalOutputName, LogicalMetric dependentMetric) {
+        LogicalMetricInfo dependentMetricInfo = dependentMetric.getLogicalMetricInfo();
+        //if no name conflict ,  return the original dependentMetric
+        if (!java.util.Objects.equals(finalOutputName, dependentMetricInfo.getName())) {
+            return dependentMetric;
+        }
+        String newName = getRenamedMetricNameWithPrefix(dependentMetricInfo.getName());
+        while (ifConflicting(newName, dependentMetric)) {
+            newName = getRenamedMetricNameWithPrefix(dependentMetricInfo.getName());
+        }
+        LogicalMetricInfo resultInfo;
+        if (dependentMetricInfo instanceof GeneratedMetricInfo) {
+            GeneratedMetricInfo generatedMetricInfo = (GeneratedMetricInfo) dependentMetricInfo;
+            resultInfo = new GeneratedMetricInfo(newName, generatedMetricInfo.getBaseMetricName());
+        } else {
+            resultInfo = new LogicalMetricInfo(newName);
+        }
+
+        return dependentMetric.withLogicalMetricInfo(resultInfo);
+    }
+
+    /**
+     * Abstract method to build new aggregation name.
+     * @param name The dependent metric name need rename.
+     * @return Renamed metric name with specified prefix
+     */
+    abstract protected String getRenamedMetricNameWithPrefix(String name);
+
+    /**
+     * Method to determine if there is name conflict.
+     * @param newName The dependent metric new name.
+     * @param dependentMetric The dependent LogicalMetric.
+     * @return Returns true if newName is already been used by the aggregations. otherwise false.
+     */
+    protected boolean ifConflicting(String newName, LogicalMetric dependentMetric) {
+        LogicalMetricInfo dependentMetricInfo = dependentMetric.getLogicalMetricInfo();
+        return java.util.Objects.equals(dependentMetricInfo.getName(), newName);
     }
 
     /**